Refactor CTAP HID implementation.
